feat: move gnoweb into his own package
