+++ gno.land/pkg/gnoweb/static/img/favicon.ico
