+++ gno.land/pkg/gnoweb/static/img/logo-square.png
