+++ gno.land/pkg/gnoweb/static/js/renderer.js
