+++ gno.land/pkg/gnoweb/static/static.go
