+++ gno.land/pkg/gnoweb/views/generic.html
