+++ gno.land/pkg/gnoweb/views/redirect.html
